+++ PipeScript.GitHubWorkflow.PSDevOps.ps1
@@ -1,5 +1,5 @@
 #requires -Module PSDevOps
-New-GitHubWorkflow -Name "Analyze, Test, Tag, and Publish" -On Push, PullRequest, Demand -Job PowerShellStaticAnalysis, TestPowerShellOnLinux, TagReleaseAndPublish, HelpOut, RunEZOut -Environment @{
+New-GitHubWorkflow -Name "Analyze, Test, Tag, and Publish" -On Push, PullRequest, Demand -Job PowerShellStaticAnalysis, TestPowerShellOnLinux, TagReleaseAndPublish, BuildPipeScript, HelpOut, RunEZOut -Environment @{
     NoCoverage = $true
 }|
     Set-Content .\.github\workflows\TestAndPublish.yml -Encoding UTF8 -PassThru