+++ Transpilers/Inline/Inline.Python.psx.ps1
@@ -1,0 +1,58 @@
+<#
+.SYNOPSIS
+    Python Inline PipeScript Transpiler.
+.DESCRIPTION
+    Transpiles Python with Inline PipeScript into Python.
+
+    Because Python does not support multiline comment blocks, PipeScript can be written inline inside of multiline string
+
+    PipeScript can be included in a Python string that starts and ends with ```{}```, for example ```"""{}"""```
+.Example
+    .> {
+       $pythonContent = @'
+"""{
+$msg = "Hello World", "Hey There", "Howdy" | Get-Random
+@"
+print("$msg")
+"@
+}"""
+'@
+        [OutputFile('.\HelloWorld.ps1.py')]$PythonContent
+    }
+
+    .> .\HelloWorld.ps1.py
+#>
+[ValidateScript({
+    $cmdInfo = $_
+    if ($cmdInfo.Source -match '\.py$') {
+        return $true
+    }
+    return $false    
+})]
+param(
+# The command information.  This will include the path to the file.
+[Parameter(Mandatory,ValueFromPipeline)]
+$CommandInfo
+)
+
+begin {
+    # We start off by declaring a number of regular expressions:
+    
+    $startComment = '(?>"""\{)'
+    $endComment   = '(?>\}""")'
+    
+    $startRegex = "(?<PSStart>${startComment})"    
+    $endRegex   = "(?<PSEnd>${endComment})"
+
+    $sourcePattern  = [Regex]::New("(?>$(
+        $startRegex, $endRegex -join ([Environment]::NewLine + '|' + [Environment]::NewLine)
+    ))", "IgnoreCase, IgnorePatternWhitespace", "00:00:05")
+}
+
+process {
+    
+    $fileInfo = $commandInfo.Source -as [IO.FileInfo]
+    $fileText      = [IO.File]::ReadAllText($fileInfo.Fullname)
+
+    .>PipeScript.Inline -SourceFile $CommandInfo.Source -SourceText $fileText -SourcePattern $sourcePattern
+}