+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.3
+* Adding Inline Python Support (#63)
+* Adding [OutputFile] transpiler (#64)
+* Building PipeScript with PipeScript (#54)
+---
+
 ## 0.0.2
 * New Transpilers:
   * .>ValidatePlatform (#58)